+++ electron/main.js
@@ -144,7 +144,8 @@ const createWindow = () => {
 			preload: path.join(__dirname, 'preload.js'),
 			contextIsolation: true,
 			nodeIntegration: false,
-			webSecurity: true
+			webSecurity: true,
+			spellcheck: false
 		}
 	});